feat: add distance, projX, projY to Position class
I added it because it is used in the test given by the teacher

+++ src/fr/epu/bicycle/GPS.java
@@ -11,7 +11,7 @@ public class GPS {
         return this.position;
     }
 
-    public void move(int x, int y) {
+    public void move(double x, double y) {
         this.position.setX(x);
         this.position.setY(y);
     }
